feat(maitake): add feature flags for both versions of `tracing`
This branch adds support for emitting both `tracing` 0.1 and `tracing`
0.2 diagnostics from Maitake. They are both controlled by feature flags,
which are additive --- both `tracing` versions _can_ be supported
simultaneously. The `tracing` configuration for tests still works as it
did previously.

The `tracing` 0.2 feature flag probably can't stay forever, if we ever
want to do a crates.io release of `maitake` before `tracing` 0.2 is
done...