docs(cordyceps): use `ptr::addr_of_mut!` instead of casts
Currently, the `cordyceps` tests and docs show the use of a `#[repr(C)]`
struct with the `Links` field as the first member of the struct, and use
a cast from `NonNull<Self>` to `NonNull<Links<Self>>` in order to
implement the `Linked::links` method. This is done to avoid the creation
of a temporary reference by dereferencing the pointer and borrowing the
`Links` field, which Stacked Borrows dislikes.

This approach is not ideal for a couple of reasons:

 * It performs a layout-dependent cast, and therefore *requires* the
   `#[repr(C)]` attribute be present to ensure soundness. This means
   that forgetting to add that attribute may result in miscompiles.
   Also, relying on struct layout makes me feel uncomfortable and we
   should try to avoid it where possible.
 * It means that a given struct may not contain multiple types of
   `Links` for different intrusive data structures. This is because the
   layout-dependent cast requires the `Links` field to be the *first*
   member of the struct, and multiple fields obviously cannot be the
   first. :)

We can also avoid the creation of a temporary reference by using the
`core::ptr::addr_of_mut!` macro to offset a pointer to a subfield of the
pointed type. This avoids the disadvantages of the cast while still
passing Miri.

Therefore, this branch updates the `cordyceps` examples, docs, and
tests, to show the use of `addr_of_mut!` rather than casting. In
addition, I've added a section to the `Linked` trait documentation
explaining why it's necessary to avoid temporary references, and
discussing why `addr_of_mut!` should be used for this purpose.

In the future, we may want to consider adding a macro (or a derive
attribute?) that generates correct `Linked` implementations using
`addr_of_mut!`, to make it easier to use intrusive structures safely...