refac(maitake): use `ptr::addr_of_mut!` instead of casting (#259)
This replaces uses of layout-dependent casts on a `#[repr(C)]` type
with the use of `core::ptr::addr_of_mut!` in the `cordyceps::Linked`
implementations for various types in `maitake`. See #258 for details on
why this is preferable.

I did *not* remove all of the layout-dependent casts in the `Task`
struct's vtable methods, so that type remains `#[repr(C)]` It would be
nice to replace those with `addr_of!`/`addr_of_mut!` as well. However,
we would likely still want this type to be `#[repr(C)]` for performance
reasons, so that the `Header`, which contains its hottest fields, is
prefetched on CPUs with spatial prefetch. However, it would be nice if
we didn't *depend* on the layout for soundness.

Since that hasn't been fully addressed in this PR, though, I've instead
added a test in `maitake::task` that asserts the address of the `Header`
field is equal to the address of the `Task` struct itself. This will
fail if we make changes that inadvertantly change the layout.