fix(maitake): fix `WaitCell` returning `Closed` when woken
There's currently a bug in `WaitCell::register_wait` where the `Closed`
error is returned if the `WaitCell` is woken during a register attempt.
I believe this was introduced in PR #223, due to a misunderstanding of
this line:

```rust
// Reset to the WAITING state by clearing everything *except*
// the closed bits (which must remain set).
let state = test_dbg!(self.fetch_and(State::CLOSED, AcqRel));
// The only valid state transition while we were parking is to
// add the CLOSED bit.
debug_assert!(
    state == actual || state == actual | State::CLOSED,
    "state changed unexpectedly while parking!"
);
```

This code does *not* set the `CLOSED` bit if it is not already set, it
just clears all other bits *except* the `CLOSED` bit. So, if the
`CLOSED` bit is not set, it remains unset, but if it is set, it is not
cleared. I think that in #223, this was misunderstood as *setting* the
closed bit, so the `closed()` error was always returned. But, the cell
may not actually have closed.

This branch changes that match arm to return the `closed()` error *if*
the closed bit was actually set by another thread, and `notified()`
otherwise.