style(alloc): fix interior mut const lint (#271)
clippy doesn't like interior mutable items in `const`s, because
mutating an instance of the `const` value will not mutate the const.
that is the *correct* behavior here, as the const is used just as an
array initializer; every time it's referenced, it *should* produce a
new value. therefore, this warning is incorrect in this case.

see https://github.com/rust-lang/rust-clippy/issues/7665