feat(util): add a checked `MaybeUninit` cell (#275)
This branch adds a `mem::CheckedMaybeUninit` type in `mycelium-util`.
This is analogous to a `MaybeUninit`, except that, in debug mode builds,
it stores a flag tracking whether the memory location has been
initialized, and asserts that the flag is set when the memory location
is accessed. In release mode builds, this flag is not present.

This is intended to help ensure that uninitialized memory is used
correctly, while avoiding the potential overhead of tracking
initialization states in release builds.

Note that this is a somewhat more limited version of `MaybeUninit`:
because it stores only a single `bool` flag to track initialization, it
cannot be used to track element-by-element initialization of arrays. It
is usable only in cases where a single `T`-typed value may be
uninitialized.

Similarly, it does not know whether a zero-initialized memory location
is valid for a given type `T`, so it does not implement
`MaybeUninit::zeroed`.