feat(maitake): add `Schedule::current_task` (#282)
This adds an initial implementation of a `Schedule::current_task` method
for cloning the currently-polled task. This implementation is somewhat
inefficient, as it has to perform an additional ref count clone/drop
cycle on every poll. This is unfortunately necessary because the current
task variable is not unset until the `TaskRef::poll` method returns, and
that method can drop the task, so we have to keep the task alive as long
as it's accessible through the scheduler (or else we risk handing out
`TaskRef`s pointed at deallocated tasks).

We can probably improve this by removing the unnecessary ref clone/drop
cycle, if we refactored how polling works a bit. If the scheduler unset
the current task variable _before_ a task is dropped, we wouldn't need
to clone the task before storing it in the current task variable...but,
this implementation works for now.

Closes #277