perf(maitake): avoid ref cycle in scheduler's current task (#283)
The implementation of `Scheduler::current_task` added in #282 is
somewhat inefficient, as it performs an additional refcount
increment/decrement cycle on every poll. By refactoring how task polling
works a little bit, we can avoid this, by polling tasks through a
non-owning reference and making the `scheduler::Core` responsible for
post-poll cleanup. This way, the `Core` owns the `TaskRef` for the
entire poll, and can store the task in its `AtomicPtr` to the
currently-polled task without having to clone an additional reference.

This is essentially the approach I proposed in [this comment][1].

While making this change, I also cleaned up the post-poll state
transition code a little bit. The new implementation is slightly
simpler.

[1]: https://github.com/hawkw/mycelium/pull/282#issuecomment-1200289502