zero segment selectors when initializing GDT
turns out the reason we were getting general protection faults was that
we would hit an `iretq` after turning on interrupts, and then try to
jump to an invalid stack segment (which i guess was set up by the
bootloader). in long mode, we have to zero the segment selector
registers, so this change does that, fixing the fault.