fix(maitake): allow getting a `JoinHandle`'s ID after completion (#298)
Currently calling the `id` method on `JoinHandle` to get the task's ID
will panic if the `JoinHandle` has already consumed the task's output,
which seems a bit problematic --- you might want to know the ID of the
task even after it has completed. Therefore, this commit changes the
`JoinHandle` type to store the task's ID when it's created, so that it
can be accessed even after the `JoinHandle`'s future has completed.

This does make the `JoinHandle` struct two words instead of one, but I
don't think that's hugely problematic.