fix(maitake): `JoinHandle` `Debug` shouldn't require `T: Debug` (#298)
Currently, the `fmt::Debug` impl for `JoinHandle`s is derived. The
derived impl will require `T: fmt::Debug` for the future's `Output`
type, even though it's stored in a `PhantomData` and the `JoinHandle`
doesn't actually own a value of that type. `JoinHandle`s probably
shouldn't require the future's `Output` to be `Debug` in order to format
a `JoinHandle`.

This commit changes the `Debug` impl to a manual implementation. It also
improves the debug formatting a bit over the derived version, such as by
including the output type's _name_.