refac(maitake): combine `wait` module into `sync` (#303)
currently, `maitake` has a `wait` module, which contains `WaitCell`,
`WaitQueue`, `WaitMap`, and (as of #301) `Semaphore`. in addition, it
has a separate `sync` module, which currently contains `Mutex` and
nothing else (and may grow a `RwLock` eventually/soon). this feels a bit
unfortunate --- the `wait` types are also synchronization primitives.

this branch moves all of `maitake::wait` into `maitake::sync`. in
addition,i fixed up the docs a little bit.

closes #302