refac(maitake): rename `WaitCell` states to match fn names (#306)
Currently, the `WaitCell` state variants are named `WAITING`, `PARKING`,
and `NOTIFYING`, but the method that puts a `WaitCell` in the `PARKING`
state is named `register_wait`, and the method that puts it in the
`NOTIFYING` state is named `wake` (rather than `notify`). This commit
renames `PARKING` and `NOTIFYING` to `REGISTERING` and `WAKING`,
respectively, to match the API methods.