chore: always use `nextest`'s `immediate-final` failure output
The `immediate-final` nextest config prints out output for failing tests
as soon as they fail, and also at the end of the run (for easy
scrollability). A lot of our tests, especially in `maitake`, will print
a bunch of logs on failure, so it's helpful to have the complete list of
failing tests printed *after* all the logs.

We used to only do this on CI. This branch changes the nextest config to
do it all the time.