feat(maitake): add owned variants of `RwLock` methods
This branch builds on the `maitake::sync::RwLock` type added in #307 by
adding `OwnedRwLockReadGuard` and `OwnedRwLockWriteGuard` types and
`read_owned`/`write_owned`/`try_read_owned`/`try_write_owned` methods to
`RwLock`, similar to the `lock_owned`/`OwnedMutexGuard` API for
`maitake::sync::Mutex`. Unlike the `RwLock{Read, Write}Guard` types,
which borrow a reference to the lock, the `Owned{Read, Write}Guard`
types hold an `Arc` clone of the `RwLock` and are therefore valid for
the `'static` lifetime.

These methods are available when the `RwLock` is wrapped in an `Arc`,
and therefore require `maitake`'s "alloc" feature flag to be enabled.