perf(maitake): wake waiters in batches outside locks  (#312)
This branch fixes some TODOs in `maitake::sync` by changing the
`WaitMap`, `WaitCell`, and `Semaphore` types so that when multiple
waiting tasks need to be woken, they are woken in batches after
releasing the lock on the queue of waiting tasks. This way, other tasks
can acquire the lock while the waiting tasks are woken, and then the
lock is reacquired if it's necessary to continue waking.

This is implemented using a new `WakeBatch` internal type, which is
essentially just an array of wakers that can be drained and reused. I
wanted to use this for my timer wheel implementation.