feat(maitake): implement cancel culture for tasks (#286)
This branch implements cancellation for spawned tasks. A task can be
cancelled through its `JoinHandle` (`JoinHandle::cancel`), or through a
`TaskRef` to it (`TaskRef::cancel`). When a task is cancelled, a bit is
set on its state field indicating that it has been canceled, and the
task is woken.

A leaf future inside a task can cancel that task by calling
`Scheduler::current_task()` to get a `Taskref` to itself, and then
canceling that `TaskRef`.

When the scheduler polls a task with the canceled bit set, it doesn't
poll the future, but instead notifies the task's `JoinHandle` and drops
its own reference to that task. Once the `JoinHandle` and all `TaskRef`s
that own reference counts for that task are dropped, the task can be
deallocated.

A task may be canceled in a state where the `Future` has completed but
the `JoinHandle` hasn't consumed the `Output` from the `Future` yet. In
that case, the `JoinHandle` still returns an error, but the error
indicates that the task was canceled after completion, and allows
recovering the future's `Output`. anyway.

Closes #264