test(maitake): compile loom tests without `cfg(loom)`
PR #286 added non-loom stub versions of `loom::model` and
`loom::alloc::track`, and changed some of the task tests to use them, so
that those tests can be compiled both with and without `cfg(loom)`.

This is nice because it means that the loom tests are still compiled
when compiling normal tests, which means that we get compiler errors for
loom tests in `rust-analyzer` output in editors, and see compilation
failures earlier. This way, we don't have to run a separate compile with
the `loom` cfg enabled just to see that a change broke the compilation
of a loom test. Also, it's nice for these tests to run normally with
non-simulated threads etc.

This branch updates all of `maitake`'s loom tests to work this way, so
that they're always compiled and run even outside of loom. In the
process, I reorganized the test directory structuring a bit, and
hopefully made it more consistent everywhere.

No functional changes.