feat(x86_64): map the kernel in the higher half (#325)
Version [0.10.13][1] of `bootloader` made it possible to configure the
start address of the dynamic range into which the kernel binary is
mapped (see rust-osdev/bootloader#229). This allows us to configure the
bootloader to boot with the kernel already mapped into the higher half
of the virtual address space, which seems nice.

[1]: https://github.com/rust-osdev/bootloader/blob/main/Changelog.md#01013--2022-09-25