test(maitake): never enable `tracing`'s `log` support (#327)
This fixes an issue where log messages in tests would sometimes be
duplicated due to `tracing`'s support for the `log` crate being enabled.
Because we (potentially) emit both a `tracing` 0.1 and `tracing` 0.2
event based on feature flags, running `maitake`'s tests with
`--all-features` results in duplicate log lines.

This is because the tests run with a `tracing` 0.2 collector set up, but
no `tracing` 0.1 collector, but the `tracing` 0.1 macros are also
enabled by `--all-features`. Therefore, the `log` support in `tracing`
0.1 is not disabled by the presence of a subscriber, and the `log`
messages are enabled...and setting up the `tracing` 0.2 collector using
`tracing-subscriber`'s `try_init` method also enables the `log` to
`tracing` adapter.

This commit changes the tests to use the lower-level
`tracing::subscriber::set_global_default` API, which doesn't also
attempt to enable a `log` -> `tracing` adapter. Now, the `log` records
will never be emitted, even if the tests are run with `--all-features`.