refac: @cratelyn-ify manual `fmt::Debug` impls
By using an exhaustive destructuring in manual `fmt::Debug` impls, we
get a compiler error if an additional field is added to a type with a
manual `Debug` impl which is not used in the `Debug` impl. This is nice,
since it helps guard against accidentally adding a field without
formatting it.

I've applied this refactoring to *most* manual `Debug` impls, with the
exception of a few that do something more weird than just manually
printing all the `Debug`-able field values.

This caught at least one case where I had forgotten to add a field to a
`Debug` impl :)