feat(pci): enumerate PCI devices (#337)
this branch adds basic support for the PCI bus to mycelium (on x86). it
adds a new `mycelium-pci` crate with general-purpose definitions of PCI
config types, and code for reading the PCI configuration space using
"configuration space access mechanism no. 1" as described [here][1].
currently, this crate works only on x86, because i didn't want to spend
a bunch of time writing abstractions to make it plug into other
architectures that support PCI, but i'll do that later.

the mycelium kernel has been updated to enumerate the PCI bus and print
what it found as part of the boot process. right now, we don't really
*do* anything of value with that information, but...future PRs will
build on top of this one to add drivers for devices on the PCI bus.

i started this branch like, three years ago, and never finished it...but
last night, i didn't sleep a lot, so today i cranked this out in a bout
of sleep-deprived coding...which feels like a bit of a pattern (remember
"flu-driven development" that got us the first booting mycelium kernel?
lmao).

thanks to @mystor for helping out with the formatting code for PCI class
enums forever ago <3

![image](https://user-images.githubusercontent.com/2796466/194780295-273b4dac-72dc-4a8d-afe4-30d44b240ef6.png)

```
[*] mycelium_kernel::arch::x86_64::pci: enumerating PCI devices...
[*]  pci: [0000:00:00] vendor=32902, device=4663
[*]  pci: class=Bridge Device
[*]  pci: [0000:01:00] vendor=32902, device=28672
[*]  pci: class=Bridge Device
[*]  pci: [0000:01:01] vendor=32902, device=28688
[*]  pci: class=Mass Storage Controller: IDE Controller: ISA compatibility mode-only controller
[*]  pci: [0000:01:03] vendor=32902, device=28947
[*]  pci: class=Bridge Device
[*]  pci: [0000:02:00] vendor=4660, device=4369
[*]  pci: class=Display Controller: VGA Compatible Controller: VGA Controller
[*]  pci: [0000:03:00] vendor=32902, device=4110
[*]  pci: class=Network Controller: Ethernet Controller
[*] mycelium_kernel::arch::x86_64::pci: found 6 PCI devices (0 bad)
```

[1]:
    https://wiki.osdev.org/Pci#Configuration_Space_Access_Mechanism_.231

Signed-off-by: Eliza Weisman <eliza@buoyant.io>
Co-authored-by: Nika Layzell <nika@thelayzells.com>