fix(bitfield): pairs still work with typed packers (#346)
Currently, the pack pair types in `mycelium-bitfield` only implement
`pack_from_src`/`pack_from_dst` if the pair doesn't have a `FromBits`
type. This is unnecessary, as it should be possible to use `FromBits`
conversion when packing a value but just move raw bits around when
packing from one bitfield to another.

This commit fixes this by adding generics to the appropriate functions
so they can still be used when the packer is generic.