feat(x86_64): implement `acpi`'s glue trait
This commit implements the `AcpiHandler` glue trait from the `acpi`
crate to allow using `acpi` in the kernel. This trait is responsible for
mapping physical memory regions that contain ACPI tables.

In this case, we know these memory regions are identity mapped, so we
don't actually have to map and unmap pages in order to implement this
trait. Instead, we just have to calculate the virtual address based on
the kernel's virtual memory offset and tell the `acpi` crate to use that
region.