feat(x86_64): read ACPI table SMP info (#347)
This commit starts on #345 by using the `acpi` crate to read information
from the ACPI tables related to SMP. It adds an `acpi::bringup_smp`
function in `arch::x86_64`, but this function doesn't *currently*
actually bring up application processors. Instead, it just reads data
from the ACPI tables and logs it as a way to check that we are, in fact,
able to find the information we need.

Additionally, I added a new `ArchInfo` struct that contains opaque
architecture-specific data from the bootloader's boot info. This is
passed into the `kernel_start` function, but the cross-platform kernel
code doesn't consume it --- instead, it's just provided so it can be
passed back into `arch_init` once we've done cross-platform setup stuff.
This felt better than sticking the RSDP addr in a global variable or
something...

This is a first step towards #345. Future commits will actually switch
to IOAPIC interrupt handling and then start the application processors.