feat(kernel): track allocator state for oopsing (#352)
This branch adds code for tracking whether or not we are currently
inside the allocator, so that we can provide this information to the
oops handler. We now print whether a panic or fault occurred inside the
allocator. In addition, we can use this to determine whether or not to
dump free lists in the oops handler, so that they are only dumped if
the allocator is implicated in an oops.

Furthermore, a new `OopsSituation` variant is added for the alloc error
handler, so that we can also dump free lists on allocation failures.