feat(hal-x86_64): add `cpu::Msr` type for accessing MSRs
This commit adds a `hal_x86_64::cpu::Msr` type for accessing x86
model-specific registers, either as raw integer values or as typed
values using the `mycelium_bitfield::FromBits` trait.

Currently, a special constructor is provided for accessing the
`IA32_APIC_BASE` MSR, which is needed for APIC interrupt handling.
Similar typed constructors for other MSRs will be added in future PRs.

This PR was factored out from #351 (which implements APIC interrupt
handling) so that it can be merged in a separate commit.