feat(x86_64): implement APIC interrupt handling (#351)
This branch adds support for APIC interrupt handling on x86_64 systems.
In order to support Survival MultiPlayer, we need to be able to handle
interrupts using the APIC rather than the PICs, because the APIC
interrupt model allows multiple CPU cores to handle interrupts.

This is part of the work towards SMP support (#345).

This is a large change to the `hal-x86_64` crate's `interrupt` module in
order to add types representing the `LocalApic` and `IoApic`, and to
change the interrupt controller abstraction to switch to handling
interrupts using APICs if the system has APICs. If APICs are not
detected, we continue to use the PICs, but cannot support SMP.

In addition, the kernel is updated to use the new HAL APIs.

Several smaller changes to other parts of mycelium have been factored
out from this PR to merge separately:

* b99fb81 refac(hal-x86_64): move `pit` and `tsc` into `time` (#356)
* 9994cd5 feat(hal-x86_64): add PIT driver (#355)
* 30431a9 fix(trace): weird indentation on new spans (#354)
* 91e97b7 feat(hal-x86_64): add `cpu::Msr` type for accessing MSRs
  (#353)

We'll probably want to make additional changes to the new interrupt code
in subsequent PRs, but this is enough to get mycelium to use the APICs,
and demonstrate that he `maitake` timer wheel still works using the
local APIC timer).