feat(hal): make `Handlers` keyboard interrupt PS/2 (#359)
The current abstraction for a generic "keyboard controller" in the HAL
`interrupt::Handlers` trait isn't great. The handler would still have to
manually read a scancode from the keyboard, making it not actually an
abstraction over various platforms, where there would be different
platform-specific ways of reading from the keyboard.

I've changed the `Handlers` trait so that that function is now
specifically a PS/2 keyboard handler (and it's renamed to
`ps2_keyboard`). It's now called with a scancode from the PS/2 keyboard
controller, so the HAL impl can actually abstract over how scancodes are
read.

We would probably have to add a different abstraction for USB HID or
other keyboard types anyway, so making this PS/2-specific seems like a
good idea, and it allows us to pass in the scancode, since the callee
now should understand that it's getting a PS/2 scancode.