feat(kernel): quick and dirty PS/2 keyboard driver (#363)
This PR adds a quick PS/2 keyboard driver that decodes keyboard
scancodes and allows async tasks to wait for keyboard input (not in the
ideal way, but it works). It currently uses a `thingbuf` and a `maitake`
`WaitCell` to buffer each keycode and wake tasks that are waiting on the
keyboard. We should probably replace the `thingbuf` with some kind of
broadcast channel that doesn't release a ring buffer slot until all
readers have consumed it, and add some kind of "reader handle" so that
we can track this. But, I haven't implemented that yet, and this works
if we assume there's only one reader task...

I also added a quick demo task that's spawned in `main`. The demo
keyboard task reads from the keyboard, buffers each character in a
`String` and waits until the user either types a newline or 80
characters, and prints that line. It even handles backspace!

Demo (note the "you typed a line!") log message:
![image](https://user-images.githubusercontent.com/2796466/197409596-53eb483e-0588-418a-9ff9-6d14576d4dab.png)

Closes #358